lesson18/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
html2text==2020.1.16
openai>=1.50.0
python-dotenv==1.0.0 